+++ src/section.h
@@ -116,7 +116,7 @@ class SectionRefs
     const_iterator begin() const { return m_entries.cbegin(); }
     const_iterator end()   const { return m_entries.cend(); }
     bool empty() const { return m_entries.empty(); }
-    int size() const   { return (int)m_entries.size(); }
+    size_t size() const { return m_entries.size(); }
 
   private:
     SectionInfoVec m_entries;
